requirements.txt
pandas
python-dotenv
tushare
duckdb
tabulate
numpy
PyYAML
